first commit
Ok guys This is my first commit to our project. The goal of this code is create a maze where the user has to move the 'P' in this case mean people through the obstacle until the user reach 'T' in this case means treasure. I program the numeral keyboard to press 8 to go up, 2 go down, 4 go left and 6 go right. X means wall the user have to find the way out and * means game over. This game is totally in C but with previous lecture we can add more feature.