Add files via upload
In this program, we will construct a class called Maze to serve as a representation of our maze game. This class includes a two-dimensional array of characters that is 10 by 10 in size and stores the contents of the maze (for example,", '#', and 'O'). In addition, we keep a record of the playerLost Boolean value, which indicates whether or not the player has died as a result of falling through a hole.
The constructor for the Maze class first positions the player at the beginning of the game, after which it populates the board with empty spots (0,0). Additionally, it generates 10-20 barriers at random and places them across the board in various locations.
The Print function does nothing more than send to the console a text representation of the current contents of the board.
The MovePlayer function accepts a character as input (for example, "w") and then moves the player in the appropriate direction based on this character. The player advances one space in the indicated direction if the input is "w," "a," "s," or "d." Other inputs have no effect. If the player attempts to move into a wall, they will not get further than where they were before.
The IsSolved function determines whether or not the player has reached the conclusion of the maze (9,9).
The IsLost function determines whether or not the player has died as a result of falling down a hole.
The player is considered to have failed to come back from their previous error and win the game once the Lost function sets the playerLost Boolean to the true value.
After constructing a Maze object in the main method, we go into a while loop to continue processing the data. This cycle will continue until either the player wins or is eliminated from the game. During each cycle of the loop, we will print out the maze, ask the player for input, move the player, and determine whether or not the player has succeeded in completing the maze. Each time it is the player's turn, there is also a remote possibility that they will be swallowed by a hole.
We will print the relevant message regardless of whether the player wins or loses the game.